Protect against single quotes matching string start
Addresses issue #10 - single quote in values break highlighting.

This causes the opening quote for a string only to be recognised when
it's preceded by a space of by the start of line, and the closing quote
to only be recognised when succeeded by a space or by the end of line.

Now you can use apostrophes in yaml files :)