fix(after syntax): use `hi def`
This is to allow colorschemes to override these highlights without having to `:hi clear`. `:hi clear`ing can be undesirable in certain situations (since it clears the highlights applied by other plugins), and an alternative to forcing this is to specify these highlights as `def` so that if a definition for a group is already in place it won't override what is there.